Collapse tracing_state.h into tracer.h
Summary: Whatever reasons for the separation originally, almost
everywhere just includes tracer.h now, making the distinction confusing.
This just collapses them to a single file.
