+++ tools/autograd/templates/variable_factories.h
@@ -6,6 +6,7 @@
 #include <c10/util/ArrayRef.h>
 #include <torch/csrc/autograd/variable.h>
 #include <torch/csrc/jit/tracer.h>
+#include <torch/csrc/jit/ir.h>
 
 #include <functional>
 #include <initializer_list>
