+++ torch/csrc/utils/python_numbers.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <torch/csrc/Exceptions.h>
-#include <torch/csrc/jit/tracing_state.h>
+#include <torch/csrc/jit/tracer.h>
 #include <torch/csrc/python_headers.h>
 #include <torch/csrc/utils/object_ptr.h>
 #include <torch/csrc/utils/tensor_numpy.h>
